refactor: replace @storybook/react with @storybook/react-vite in remaining stories (batch 3)
