+++ apps/storybook/stories/dyn-select.stories.tsx
@@ -1,4 +1,4 @@
-import type { Meta, StoryObj } from '@storybook/react'
+import type { Meta, StoryObj } from '@storybook/react-vite'
 import { userEvent, within, expect } from '@storybook/test'
 import { DynSelect } from '@dynui/core'
 import type { DynSelectRef } from '@dynui/core'
