+++ stories/DynListView.stories.tsx
@@ -1,4 +1,4 @@
-import type { Meta, StoryObj } from '@storybook/react'
+import type { Meta, StoryObj } from '@storybook/react-vite'
 import { DynListView } from '../src/ui/dyn-listview'
 
 const meta: Meta<typeof DynListView> = {
