+++ stories/DynRadioGroup.stories.tsx
@@ -1,4 +1,4 @@
-import type { Meta, StoryObj } from '@storybook/react'
+import type { Meta, StoryObj } from '@storybook/react-vite'
 import { DynRadioGroup } from '../src/ui/dyn-radio'
 
 const meta: Meta<typeof DynRadioGroup> = {
